dom/bindings/test/test_bug852846.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=852846
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 852846</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Bug 852846 **/
  var elem = document.createElement("div");
  is(elem.style.color, "", "Shouldn't have color set on HTML element")
  elem.style = "color: green";
  is(elem.style.color, "green", "Should have color set on HTML element")

  elem = document.createElementNS("http://www.w3.org/2000/svg", "svg");
  is(elem.style.color, "", "Shouldn't have color set on SVG element")
  elem.style = "color: green";
  is(elem.style.color, "green", "Should have color set on SVG element")
  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=852846">Mozilla Bug 852846</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
